+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 0.9.1 - 2018-09-12
+### Added
+- Lightswitch settings on the field type for sorting and allowing a blank state
+- Config settings for setting the default state of these lightswitches
+- Fancy new icon 
+
+### Changed
+- Switched to Choices.js for select enhancement
+
+### Fixed
+- Dropdowns now work inside of modals (no more clipping by overflow)
+
 ## 0.9.0 - 2018-05-08
 ### Added
 - Initial release